+++ test/spec/setvalue.spec.js
@@ -386,8 +386,7 @@ describe( 'setvalue actions to populate a value if another value changes', () =>
         expect( myAgeChangedView.textContent ).toEqual( '' );
         expect( myAgeChangedModel.textContent ).toEqual( '' );
 
-        form.input.setVal( myAgeView, '11', null );
-        myAgeView.dispatchEvent( events.Change() );
+        form.input.setVal( myAgeView, '11', events.Change()  );
 
         setTimeout( () => {
             //expect( myAgeChangedView.textContent ).toEqual( '6' );
@@ -406,8 +405,7 @@ describe( 'setvalue actions to populate a value if another value changes', () =>
         expect( ageChangedsView.map( el => el.textContent ) ).toEqual( [ '', '' ] );
         expect( ageChangedsModel.map( el => el.textContent ) ).toEqual( [ '', '' ] );
 
-        form.input.setVal( agesView[ 0 ], '22', null );
-        agesView[ 0 ].dispatchEvent( events.Change() );
+        form.input.setVal( agesView[ 0 ], '22', events.Change()  );
 
         setTimeout( () => {
             //expect( ageChangedsView.map( el => el.textContent )).toEqual( [ 'Age changed!', '' ] );
@@ -428,8 +426,7 @@ describe( 'setvalue actions to populate a value if another value changes', () =>
         const dModel = form.model.xml.querySelector( 'd' );
         const eModel = form.model.xml.querySelector( 'e' );
 
-        form.input.setVal( dView, '3030', null );
-        dView.dispatchEvent( events.Change() );
+        form.input.setVal( dView, '3030', events.Change() );
 
         expect( aView.textContent ).toEqual( '' );
         expect( bModel.textContent ).toEqual( '' );
@@ -438,8 +435,7 @@ describe( 'setvalue actions to populate a value if another value changes', () =>
         expect( dModel.textContent ).toEqual( '3030' );
         expect( eModel.textContent ).toEqual( 'default' );
 
-        form.input.setVal( aView, '11', null );
-        aView.dispatchEvent( events.Change() );
+        form.input.setVal( aView, '11', events.Change() );
 
         setTimeout( () => {
             expect( bModel.textContent ).toEqual( '2' );
@@ -450,8 +446,19 @@ describe( 'setvalue actions to populate a value if another value changes', () =>
             expect( eModel.textContent ).toEqual( '' );
             done();
         }, 100 );
+    } );
 
+    it( 'works if the setvalue trigger is a calculation', () => {
+        const form = loadForm( 'setvalue-triggered-by-calc.xml' );
+        form.init();
+        const a = form.input.find( '/data/a', 0 );
 
+        form.input.setVal( a, '1', events.Change() );
+
+        // check calculated value
+        expect( form.model.xml.querySelector( 'a_copy' ).textContent ).toEqual( '1' );
+        // check setvalue-changed value
+        expect( form.model.xml.querySelector( 'triggered' ).textContent ).toEqual( '11' );
     } );
 
 } );